Clarify global-namespace explanation
I had trouble understanding this passage, but a peek in the documentation
explained it.  It's not about "execution" – the execution of the function is
unaffected – it's about name resolution. And it's also not really about
"definition".

While I was at it, I renamed "method" to function, as these are not methods and
methods aren't really affected by namespaces (they're already namespaced by a
class). Also copyedited a bit.